use newtype instead of enum
To ensure backwards compatibility, don't hard-code the list of supported
PoRep and PoSt proof partitions into an enum. Don't hard-code the list
of supported SectorSizes into an enum either. Instead, accept whatever
values come in through the FFI and plumb them through to the internal.rs
functions. If the callee is asked to verify a proof that they don't have
a VK for, we should be able to go get the VK. This commit doesn't solve
the problem of a missing VK.